Fix Tensor:index for some tensors with the first dimension of size 1
Previously, the code assumed that for contiguous tensors, stride[0]
is the product of size[1] * size[2] * ... size[nDim-1]. This is not
necessarily the case if the size of the first dimension is 1.

For example, contiguous 3x2x7 and 4x2x7 tensors must have stride 14x7x1, but
a 1x2x7 tensor may have any stride in the first dimension (e.g. 99x7x1).

A correct way to determine the effective "row size" for Tensor:index
is to take the nElement / size[0].