report offending blob name when blob in wrong device scope
Summary:
Another recurrent problem is that some blob is in CPU scope while operator expects CUDA scope (or other way round).
The exception is only partially helpful, as it tells the operator but not the offending blob name. This diff adds the blob name
to the exception message, helping debug.

Reviewed By: prigoyal

Differential Revision: D4208584

fbshipit-source-id: 5aeac5c3efeed8d6c995bea166ed534855007945