make ReshapeOp work with CUDA
Summary:
It was not just enough to register ReshapeOp for CUDA, since it does memory copy to/from tensors. This happened in two places: when assigning shape from a shape blob and when outputing a shape tensor.

Also changed the resizeoptest to use CUDA when available (this test was done before hypothesis-tests, so I had to do this manually)

Differential Revision: D4217342

fbshipit-source-id: 61761bac015f3731cf480ccef2563e9c80e0f4aa