Add optional gradient on weights for (Sparse)LengthsWeightedSum
Summary:
It ended up much messier than originally expected. Maybe we should have just hardcode it, but I've tried to be "generic" so far at expense of code readability.

The main issue is that for weights computation we need access to original embedding matrix and in sparse case we need to relookup the embeddings to do the dot product with output grads.

Thus I'm making weight grad computation optional, controlled by a flag and it triggers invocation of a different backward op that produces both grads at the same time.

So far it's implemented only for 'Lengths' version. It'd be straightforward to implement (Un)SortedSegment versions but I haven't done that yet.

Reviewed By: kennyhorror

Differential Revision: D4388215

fbshipit-source-id: 23132ab7daa1f5eec49233f802af1fe75b469c2b