Fix race in FileStoreHandler
Summary:
It was possible for a set and a get to race such that the get
would return an empty string, if the file for the key was created but
not yet written to. This change updates the FileStoreHandler to first
write to a temporary file and then atomically rename(2) the file to
its final path. This removes the described race condition.

This change also replaces the poor filename generation routine with
using the 128-bit MurmurHash of a key.

Differential Revision: D4502154

fbshipit-source-id: f2abc78b8bad68c06ad2f18a078935826e431f7a