refactor LSTM test to avoid copy pasta, improve speed 1.5x and provide better coverage
Summary:
This is getting too messy again. So cleaning up it even more. One thing I added here - not calling random to generate the input sequence. Ideally we do this for all other inputs. This was reported to be an issue when hypothesis finds bad examples - it can make it run very long.

Also I tunned ranges a bit so test finishes faster. On my devgpu test the whole test took 600 before and now is 39 seconds.

One more important thing - we want to test all combinations of things that are in the for loop. While things provided by hypothesis are just random tensor inputs.

Differential Revision: D4902956

fbshipit-source-id: ceb02d6761406b3192101d3b255abe90b2866770