Make Context::FinishDeviceComputation throw instead of FATAL
Summary:
We shouldn't LOG(FATAL) in Caffe2 code under any conditions as it's a library.

The case where it failed was a bug in SparseAdaGrad that failed on empty input trying to launch 0-sized CUDA kernel.

Also, the trend for C2 core is in moving from bool to exceptions, so I just moved CAFFE_ENFORCE directly into FinishDeviceComputation. Most of the use cases were already doing that or ignoring the output (bad!).

Reviewed By: akyrola

Differential Revision: D5495913

fbshipit-source-id: 66f382369417a262da69d54470f720e7d04a5cdf