adding operator lp_norm to support calculating l1 norm and l2 norm
Summary: Implement operators LpNorm, which is to calculate the Lp norm of a tensor for regularization(p=1or 2) . Currently, there are only operator L1Distance to calculate the l1 distance of two same-shape tenors. We want to make it take only one input and output the l1 loss. We would do the same for l2 loss. We also plan to implement l_{p,q} loss, but have not decided which p and q to take.

Reviewed By: xianjiec

Differential Revision: D5460051

fbshipit-source-id: d67a38fbc94afa52de26d4a53e4d2b7df3c50b6a