Make Gloo depend on Caffe2 NCCL build
Summary:
If Caffe2 used the packaged NCCL version then the Gloo build will try
to use it as well. To make sure the NCCL build has completed we need
to add an explicit dependency between the two.

Another subtle change here is that we add the PROJECT_BINARY_DIR to
the include path, since that is where the generated <gloo/config.h>
resides. Without this path Caffe2 includes the empty config.h from the
source tree.
Closes https://github.com/caffe2/caffe2/pull/1170

Differential Revision: D5779002

Pulled By: pietern

fbshipit-source-id: 9bc0d41f01a9b0f023d71bc4dee128a77eec1712