Make docker image built on Jenkins usable out of the box
Summary:
The install prefix we use in our builds is /usr/local/caffe2. This is
not standard, so in order to load caffe2 from Python, the Python
interpreter must know where to find it. In a post-build section in the
Jenkins build script we know add a symlink to Python's dist-packages
directory and instruct the loader to look in /usr/local/caffe2/lib.
Together, these tricks make it usable out of the box.
Closes https://github.com/caffe2/caffe2/pull/1617

Differential Revision: D6572322

Pulled By: pietern

fbshipit-source-id: c37b789a0d0babbb1110f991318c6b75fe351c0e