Allow TypeMeta hold non-default-constructible types (#8349)
Necessary for Tensor detemplatization (D8121878) - now tensor won't have default constructor (as we don't know the device).

Thus this diff makes TypeMeta be constructible with non-default-constructible types in which case ctor() is non-null but always throws.

It's dangerous however as we won't catch potential type errors at compile time. Luckily - the only place where ctor() is used is in Blob and Tensor which have templated wrappers there (GetMutable and mutable_data respectively). We can just enforce the necessary type requirements there explicitly as a static_assert.

It also changes the failure behavior to be throw() instead of abort(). Aborting the process is not cool for the library :)