Fix out-of-boundary conversion issue (#11338)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/11338

The `min_` and `max_` value of the filler is in `double` format but when we are filling a specific type of tensor, their value can exceed the type limits, resulting in crash. This diff checks the type limits first and if `min_`/`max_` is out of the limits, it will clip it.

Reviewed By: highker

Differential Revision: D9684455

fbshipit-source-id: 6da98a03c57f3296abaddc7c5cfc1c836c611eb0