Remove OMP_NUM_THREADS and MKL_NUM_THREADS settings from docker images (#12836)
Summary:
`OMP_NUM_THREADS` and `MKL_NUM_THREADS` are set to 4 by default in the docker images, which causes `nproc` to only show 4 cores in the docker containers by default, and building PyTorch is slow in this default case. We likely don't need these two flags to be set, and this PR tests that hypothesis.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/12836

Differential Revision: D10468218

Pulled By: yf225

fbshipit-source-id: 7a57962c962e162a8d97f730626825aa1e371c7f