CircleCI: fix test timeout by running CPU build and test on different machines (#13284)
Summary:
It seems that we can fix the test timeout issue by running CPU build and test on different machines (I manually ran this patch through the CI 50 times to confirm this). The actual reason of timeout is still unknown, but I suspect it has to do with memory / disk space.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/13284

Differential Revision: D12840371

Pulled By: yf225

fbshipit-source-id: af326f0358355602ee458696c3ffb325922e5289