Fix Trace inlining of graphs with optional inputs (#22686)
Summary:
Previously in tracing when we called a script function we would inline the graph and set the graph inputs equal to the types the graph was invoked with.

This breaks for optional arguments invoked with None since we rely on None being set to Optional[T] in schema matching.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/22686

Differential Revision: D16186372

Pulled By: eellison

fbshipit-source-id: e25c807c63527bf442eb8b31122d50689c7822f5