fix test_arange and bump ort ci version (#26320)
Summary:
It appears to be a bug with test_arange, which wasn't revealed with older version of onnxruntime.

TLDR. The test tries to update exported onnx model to accept dynamic sized input, however it is written incorrectly such that the exported model input is still fixed sized. Meanwhile, the version of ort in CI doesn't validate if model input size matches with input data, so this error was not found.

Affecting ci in https://github.com/pytorch/pytorch/pull/25797
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26320

Reviewed By: hl475

Differential Revision: D17406442

Pulled By: houseroad

fbshipit-source-id: a09ad4b925ccbed0b71342f5aaa7878e1c4a5a2d