use unordered_set instead of vector for futureTimeouts key in (#31813)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31813

Closes https://github.com/pytorch/pytorch/issues/31804. We were using
an `std::vector` for the key for a map that keeps track of futures to mark them
if they timeout, but we can instead use an `unordered_set`. This results in a
faster lookup in the code block where we remove futureIDs from this set when
they complete successfully. Previously we were finding them via a linear
`std::find`. Switching it to a constant time find will help performance in the
case where a large number of futures are scheduled to time out at the same
time, or if there is no timeout enforced.

To benchmark a rough perf improvement, I created 50k futures with the same
timeout. Before this PR, the lookup `std::find(futuresAtTime.begin(),
futuresAtTime.end(), id)` took ~200us, now it takes 1us.
ghstack-source-id: 96251355

Test Plan: Unit tests pass.

Differential Revision: D19269798

fbshipit-source-id: 1a0fa84a478ee27a16ab0b9fa6f5413b065a663e