ecr_gc: Iterate through all tags, reduce prints (#42492)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42492

There's a potential for multiple tags to be created for the same digest
so we should iterate through all potential tags so that we're not
deleting digests that are associated with tags that we actually want.

Also, reduced the number of prints in this script to only the absolutely
necessary prints. (i.e. only the deleted images)

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>

Test Plan: Imported from OSS

Reviewed By: ezyang

Differential Revision: D22909248

Pulled By: seemethere

fbshipit-source-id: 7f2e540d133485ed6464e413b01ef67aa73df432