Use storage.cpu() for moving storage to CPU in serialization. (#46028)
Summary:
As reported in https://github.com/pytorch/pytorch/issues/46020, something seems to go wrong with the storage._write_file method used with a BytesIO and a GPU buffer.
Given that we were going to create the intermediate buffer (currently via BytesIO) anyway, we might as well use storage.cpu() to move the storage to the CPU. This appears to work better.

This is a hot fix, further investigation is highly desirable. In particular, I don't have a reproducing test to show.

Fixes https://github.com/pytorch/pytorch/issues/46020

Pull Request resolved: https://github.com/pytorch/pytorch/pull/46028

Reviewed By: bdhirsh

Differential Revision: D24194370

Pulled By: gchanan

fbshipit-source-id: 99d463c4accb4f1764dfee42d7dc98e7040e9ed3