[codemod] Enforce proper use of emplacy functions
Summary: The goal of this diff is enforce proper use of "emplacy" functions. In each case, this saves at worst a move constructor call, and at best a full copy of the object (in the case of a constructor call where the object does not have a move constructor).

Test Plan: CI.

Reviewed By: marksantaniello

Differential Revision: D27888714

fbshipit-source-id: 235d0b31066463588c7e4ab86e132c430a352500