Don't bind Android NNAPI on Apple platforms (#58323)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58323

Currently there is no way to run NNAPI on Apple platforms.
Disabling the binding with the preprocessor makes this easier
to enable NNAPI in the internal build without affecting iOS size.

This should be reverted soon and migrated to selective build.

Test Plan: Build Size Bot on later diff.

Reviewed By: axitkhurana

Differential Revision: D28435179

fbshipit-source-id: 040eeb74532752630d329b15d5f95c538c2e3f9e