[torch][package] Fix importlib.resources.path for python <3.8.8 (#58718)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58718

`PackageImporter` does not populate `module.__spec__.origin`, which causes an
unhandled `Exception` to be raised when using `importlib.resources.path` to get
a path to a binary file resource in the package in python <3.8.6.

This commit fixes this issue by setting `module.__spec__.origin` to
"<package_importer>". The actual value is not important as far as I can tell;
the simple fact that it is not `None` allows `importlib` to avoid raising an
`Exception` in `importlib.resources.path`.

Test Plan:
This commit adds a unit test to `test_resources.py` that tests that
`importlib.resources.path` can be used within a package.

Reviewed By: suo

Differential Revision: D28589117

fbshipit-source-id: 870d606a30fce6884ae48b03ff71c0864e4b325f