Send test reports to S3 (#61071)
Summary:
This sends the test reports zip to S3 in addition to the GitHub artifact store. This makes it easier to query in the PR HUD since we don't have to deal with the GitHub API's rate limits / download speeds. The impact on S3 storage should be minimal since it's only 500 KB or so per run.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/61071

Reviewed By: nikithamalgifb

Differential Revision: D29498941

Pulled By: driazati

fbshipit-source-id: 74bfbe7fa7d1d97fd8a6938c98dfe0caff0ab6eb