Add bazel actions workflow (#61039)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61039

- Added a new template for bazel GH Actions workflow
- Simplified the workflow based on malfet's suggestion by combining build and test jobs into one as we only run a small subset of tests for bazel
- Tested the run to make sure it succeeds
- Build step takes 4 minutes, test step takes 7 minutes

The downside of this approach is that I duplicated some of the jobs in a new template file.  Alternative solution would be to use something like this https://jinja.palletsprojects.com/en/3.0.x/templates/#template-inheritance, however, that is better to be done in a separate PR as linux and windows workflows would need to be changed. Another solution is to use a bunch of if else statements in a linux workflow template to accommodate bazel build as part of it, but this seems not as clean as template inheritance with jinja.

Here is a link to the latest bazel run with this change https://github.com/pytorch/pytorch/actions/runs/1004656584

Test Plan: Imported from OSS

Reviewed By: malfet

Differential Revision: D29562260

Pulled By: rsemenov

fbshipit-source-id: a7d7d3a0b8092f52929fb109820bfad4574f5602