[vulkan] Remove unused code for Winograd convolutions (#80699)
## Context

Code for Winograd convolutions is currently unused. Unfortunately, it only applies to 3x3 convolutions and it does not outperform the sliding window algorithm on the input sizes typically used in practice.

This PR remove code for Winograd convolutions as we don't foresee it being used in the future.

Differential Revision: [D36493834](https://our.internmc.facebook.com/intern/diff/D36493834/)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80699
Approved by: https://github.com/kimishpatel