[ao] Add line plot visualization capability to ModelReportVisualizer (#81974)
Summary: This adds the capability to visualize the line plot in the
ModelReportVisualizer. You can visualize line plots of single feature,
and this feature can either be a per-tensor or per-channel feature. If
the feature is per tensor, then the idx of the module is printed as the
x axis and the values of the feature as the y. If the feature is per
channel, then **an** (the first one) idx of the module will the value on
the x axis and the corresponding feature val in the y axis, and there
will be a seperate line for each channel, and a legend denoting which
line belongs to which channel.

Expected Usage
```
mod_rep_visualizer.generate_plot_visualization(<feature_name>)
```

You can also filter the modules so that only modules with a certain
substring will have their features represented in the plot.

> ** This is intended to be used in a `.ipynb` style notebook**

The tests for this were just visual inspection for two reasons:
    1.) This method does not return anything, it just generates the
    visualization plot
    2.) All the data to create the plot visualization is gotten from
    `generate_filtered_tables` which is already tested, so testing all that
    for this again would be redundant.

Example Image outputs are pasted below in the PR thread.

Test Plan: Visual Test

Reviewers:

Subscribers:

Tasks:

Tags:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81974
Approved by: https://github.com/jerryzh168