[Edge] Add an option to avoid adding base ops to static op library (#84360)
Summary: We use a static op library in a test for PyTorch C++ usages, but don't want to introduce all base ops. Because the goal is to check if a given model can run on the exact op collection (i.e., fbios ops, fb4a ops), and these base ops are not present in real apps. So add an option to disable this feature.

Test Plan: Build. Expect no change to existing targets.

Differential Revision: D39164021

Pull Request resolved: https://github.com/pytorch/pytorch/pull/84360
Approved by: https://github.com/kit1980