Disallow saved tensor hooks in functorch transforms (#85972)
Technically they may only be a problem with the grad transform. Though
the branch cut is soon, this is the more conservative change, it also
lets us disable checkpointing for functorch (which definitely doesn't
work with all transforms) and not a lot of people use saved tensor hooks
with functorch (I discovered this while testing).

Test Plan:
- new tests

Differential Revision: [D39970934](https://our.internmc.facebook.com/intern/diff/D39970934)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85972
Approved by: https://github.com/samdow