[ONNX] Fix potential flaky test in test_verification.py (#92105)
Very low probability, but it is possible to have all values positive throughout the
execution of this test model. The test tries to fake an incorrect export by replacing
relu's output with its input. However, the behavior of the model is the same when
values are all positive. Hence leading to false test failure.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/92105
Approved by: https://github.com/titaiwangms