Make ShapeEnv deepcopy'able (#93403)
We sometimes put ShapeEnv on GraphModule, and code in our testing
utils assume that you can deepcopy a GraphModule, so it's good
for ShapeEnv to be deepcopy'able too.  This is done by making the
TLS module-wide rather than per-ShapeEnv.  We never really have
multiple ShapeEnv so this is a good trade.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>
Pull Request resolved: https://github.com/pytorch/pytorch/pull/93403
Approved by: https://github.com/jbschlosser