Manually resolve download.pytorch.org to IPv4 (#100436)
This is an attempt to address https://github.com/pytorch/pytorch/issues/100400 by using only IPV4 when accessing the domain.

I kind of want to ignore AAAA records from DNS instead, but couldn't find an easy way to do so.  https://www.cloudflare.com/learning/dns/dns-records/dns-aaaa-record doc mentions this

```
Like A records, AAAA records enable client devices to learn the IP address for a domain name. The client device can then connect with and load the website.

AAAA records are only used when a domain has an IPv6 address in addition to an IPv4 address, and when the client device in question is configured to use IPv6
```

Pull Request resolved: https://github.com/pytorch/pytorch/pull/100436
Approved by: https://github.com/malfet