Update initializer path for ONNXProgram.save due to onnx.checker limitation (#117294)
According to https://github.com/onnx/onnx/blob/main/docs/ExternalData.md#large-models-2gb when initializers are larger than 2GB, `onnx.checker` requires the model to be in the same directory as the initializer.

Although not strictly necessary for the export and model save to succeed, it is desirable to have the `onnx.checker` to succeed when validation the resulting large model.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117294
Approved by: https://github.com/BowenBao