[TD] llm retrieval to not use bash -l {0} (#133464)
https://github.com/pytorch/pytorch/pull/129720 swapped the action used to setup miniconda from [conda incubator](https://github.com/conda-incubator/setup-miniconda) to the [custom action](https://github.com/pytorch/test-infra/blob/2aba8f107aca99710cca5cd97b6c16500eb808e2/.github/actions/setup-miniconda/action.yml#L1) we have in test-infra that comes with caching.

The original miniconda [relies on bash profiles](https://github.com/conda-incubator/setup-miniconda/blob/e5293c8fd23bf30f6ad37a6281c8a924c8938440/README.md?plain=1#L746) to set the environment variables needed to run conda, but the test infra version relies on the user using the env vars that are set during the step.

This PR changes the job to not use `bash -l {0}` to see if not activating bash profile has an effect on the run.  Unfortunately this failure happens rarely on main so I'm not sure I will be able see if this has an effect.  On the plus side, changing this doesn't seem to have a negative effect on the job, so it should be a noop at worst.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/133464
Approved by: https://github.com/kit1980