ci: Remove --progress-bar fallback for pip (#140189)
All versions of pip that we currently support should have this flag so removing this should essentially be a no-op.

Also put the actual command into a variable so we only have to change it once next time instead of changing it in 3 places.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/140189
Approved by: https://github.com/kit1980, https://github.com/malfet