[easy] Log bypass reasons if we're unable to serialize or deserialize a saved graph (#141911)
When we fail to deserialize/serialize a graph, we should alert and log it somewhere so that it's debuggable.

This can happen in OSS if we use view_replay and encounter an output that requires functional tensor to be serialized to work.

Differential Revision: [D66669993](https://our.internmc.facebook.com/intern/diff/D66669993/)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141911
Approved by: https://github.com/oulgen, https://github.com/ezyang